+++ src/App.js
@@ -1,66 +1,14 @@
 import * as S from './App.styles';
 import AppRoutes from './routes';
 import { GlobalStyle } from './createGlobalStyle';
-import { useState, useEffect, useReducer } from "react";
-import { getTrackById, getTracks} from "./api";
-// import { useBeforeRender } from './utils';
 import { UserContext, UserDispatchContext, reducer } from './contex';
+import { useReducer } from 'react';
 
 
 function App() {
-  // const [user, setUser] = useState(null);
-
-  // useBeforeRender(() => {
-  //   const registeredUser = localStorage.getItem('user');
-  //   if (registeredUser) {
-  //     setUser(JSON.parse(registeredUser));
-  //   }
-  // }, []);
-
-  // const handleLogin = () => {
-  //   const newUser = { login: 'taradam' };
-  //   setUser(newUser);
-  //   localStorage.setItem('user', JSON.stringify(newUser));
-  // };
-
-  // const handleLogout = () => {
-  //   setUser(null);
-  //   localStorage.removeItem('user');
-  // };
-
-  const [isLoading, setIsLoading] = useState(true);
-
-  const [tracks, setTracks] = useState([])
-  const [addTracksError, setAddTracksError] = useState(null)
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setIsLoading(false);
-      try {
-        getTracks().then((t) => 
-        {setTracks(t)}
-      );
-      } catch (error) {
-        setAddTracksError(error.message)
-      }
-
-    }, 5000);
-
-    return () => clearTimeout(timer);
-  }, []);
-
-  const [currentTrack, setCurrentTrack] = useState(null);
-  const turnOnTrack = (trackId) => {
-      getTrackById(trackId)
-        .then(trackData => {
-          setCurrentTrack(trackData);
-          console.log(trackData);
-        })
-    }
     const userState = {
       userName: JSON.parse(localStorage.getItem('user')) || ''
     };
-    
     const [state, dispatch] = useReducer(reducer, userState)    
  
   return (
@@ -68,13 +16,7 @@ function App() {
     <UserDispatchContext.Provider value={dispatch}>
     <S.App>
         <GlobalStyle/>
-          <AppRoutes 
-          tracks={tracks} 
-          isLoading={isLoading}
-          currentTrack={currentTrack} 
-          turnOnTrack={turnOnTrack} 
-          addTracksError={addTracksError}
-        />
+          <AppRoutes />
     </S.App>
     </UserDispatchContext.Provider>
     </UserContext.Provider>
