+++ src/contex.js
@@ -22,8 +22,8 @@ export const reducer = (state, action) => {
     }
   }
 
- export const checkUserExists = (username) => {
-    const users = useContext(UserContext);
-    const userExists = users.find((user) => user.username === username);
-    return !!userExists;
-  };
+//  export const checkUserExists = (username) => {
+//     const users = useContext(UserContext);
+//     const userExists = users.find((user) => user.username === username);
+//     return !!userExists;
+//   };
